- More minor code cleanups.
git-svn-id: http://svn.code.sf.net/p/qjackctl/code_attic/trunk@420 5e98d84e-b994-457e-aaaa-8ae0a2fdc4f7

+++ src/qjackctlMainForm.h
@@ -153,7 +153,7 @@ protected:
 
 	void updateXrunCount();
 
-	QString formatTime(double secs) const;
+	QString formatTime(float secs) const;
 	QString formatElapsedTime(int iStatusItem, const QTime& t, bool bElapsed) const;
 	void updateElapsedTimes();
